Break large home page component into multiple
